Add refractive index calculation from Liebe et al. (1991)
The results from this calculation do not significantly differ from those
of Ray (1972) over the temps/wavelengths of interest; therefore, this
calculation has not been implemented as an option. However, since I had
to code it up to test, it seemed a shame to just throw it away. Hence,
it lives in here now as a function.

Really, a bunch of refactoring is necessary to make it easier to utilize
this calculation if so desired.